pyproject.toml
[project]
name = "forge"
version = "0.1.0"
description = "基于红黄蓝白(RYBW)叠色原理的多色3D打印文件生成器"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "lib3mf>=2.4.1.post1",
    "numba>=0.60.0",
    "numpy>=2.2.6",
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "pyside6>=6.10.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "typing-extensions>=4.15.0",
]

[project.scripts]
forge = "forge.main:main"

[tool.uv]
dev-dependencies = ["pytest>=7.0.0", "pytest-qt>=4.2.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/forge"]
